+++ demo/build.gradle
@@ -35,4 +35,6 @@ dependencies {
     compile("org.mybatis.spring.boot:mybatis-spring-boot-starter:2.0.0")
     compile("org.mybatis:mybatis-spring:2.0.0")
     compile("org.mybatis:mybatis:3.5.0")
+
+    compile("org.springframework.boot:spring-boot-starter-security")
 }
