--- demo/src/main/java/hello/CityMapper.java
@@ -1,10 +1,0 @@
-package hello;
-
-import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Param;
-
-
-@Mapper
-public interface CityMapper {
-    City findByState(@Param("state") String state);
-}
