+++ demo/src/main/java/hello/Member.java
@@ -1,0 +1,34 @@
+package hello;
+ 
+import java.util.Collection;
+ 
+public class Member {
+ 
+    private Long id;
+    private String name;
+    private String password;
+
+    public Long getId() {
+        return this.id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return this.name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+ 
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
