+++ demo/src/main/resources/data.sql
@@ -1,2 +1,4 @@
-INSERT INTO city (name, state, country) VALUES ('San Francisco', 'CA', 'US'); 
+-- INSERT INTO city (name, state, country) VALUES ('San Francisco', 'CA', 'US'); 
+
+INSERT INTO users (name, password) VALUES ('user1', 'password');
