+++ demo/src/main/resources/hello/UserMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="hello.CityMapper">
-  <select id="findByState" resultType="hello.City">
-    SELECT * FROM city WHERE state = #{state}
+<mapper namespace="hello.UserMapper">
+  <select id="findById" resultType="hello.Member">
+    SELECT * FROM users WHERE id = #{id}
   </select>
 </mapper>
