+++ demo/src/main/resources/schema.sql
@@ -1,7 +1,17 @@
-DROP TABLE IF EXISTS city;
-CREATE TABLE city (
+-- DROP TABLE IF EXISTS city;
+-- CREATE TABLE city (
+--     id INT PRIMARY KEY AUTO_INCREMENT,
+--     name VARCHAR(255),
+--     state VARCHAR(255),
+--     country VARCHAR(255)
+-- );
+DROP TABLE IF EXISTS users;
+CREATE TABLE users (
     id INT PRIMARY KEY AUTO_INCREMENT,
-    name VARCHAR(255),
-    state VARCHAR(255),
-    country VARCHAR(255)
-); 
+    name VARCHAR(255) NULL DEFAULT NULL,
+    password VARCHAR(500) NULL DEFAULT NULL
+)
+COLLATE='utf8_general_ci'
+ENGINE=InnoDB
+;
+
