+++ demo/src/main/webapp/WEB-INF/jsp/account_detail.jsp
@@ -1,0 +1,12 @@
+<%@ page language="java" contentType="text/html; charset=utf-8" pageEncoding="utf-8" isELIgnored="false"%>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
+
+<html>
+<head>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+    <title>account_detail</title>
+</head>
+<body>
+    account_detail
+</body>
+</html>
